+++ test/conftest.py
@@ -1,0 +1,7 @@
+import pytest
+
+
+@pytest.fixture
+def saiga_antelope():
+    return 'Saiga antelope have distinctive bulbous noses. These help filter out dust and regulate their' \
+           'body temperature in the harsh climates of Central Asia.'